Create top level package for query datatypes.
