+++ src/logicblocks/event/projection/__init__.py
@@ -4,46 +4,10 @@ from .projector import (
 )
 from .projector import Projector as Projector
 from .store import (
-    Clause as Clause,
-)
-from .store import (
-    FilterClause as FilterClause,
-)
-from .store import (
     InMemoryProjectionStorageAdapter as InMemoryProjectionStorageAdapter,
-)
-from .store import (
-    KeySetPagingClause as KeySetPagingClause,
-)
-from .store import (
-    Lookup as Lookup,
-)
-from .store import (
-    OffsetPagingClause as OffsetPagingClause,
-)
-from .store import (
-    Operator as Operator,
-)
-from .store import (
-    Path as Path,
 )
 from .store import (
     PostgresProjectionStorageAdapter as PostgresProjectionStorageAdapter,
 )
 from .store import ProjectionStorageAdapter as ProjectionStorageAdapter
 from .store import ProjectionStore as ProjectionStore
-from .store import (
-    Query as Query,
-)
-from .store import (
-    Search as Search,
-)
-from .store import (
-    SortClause as SortClause,
-)
-from .store import (
-    SortField as SortField,
-)
-from .store import (
-    SortOrder as SortOrder,
-)
