+++ src/logicblocks/event/projection/store/__init__.py
@@ -5,43 +5,4 @@ from .adapters import (
     PostgresProjectionStorageAdapter as PostgresProjectionStorageAdapter,
 )
 from .adapters import ProjectionStorageAdapter as ProjectionStorageAdapter
-from .query import (
-    Clause as Clause,
-)
-from .query import (
-    FilterClause as FilterClause,
-)
-from .query import (
-    KeySetPagingClause as KeySetPagingClause,
-)
-from .query import (
-    Lookup as Lookup,
-)
-from .query import (
-    OffsetPagingClause as OffsetPagingClause,
-)
-from .query import (
-    Operator as Operator,
-)
-from .query import (
-    Path as Path,
-)
-from .query import (
-    Query as Query,
-)
-from .query import (
-    Search as Search,
-)
-from .query import (
-    Similarity as Similarity,
-)
-from .query import (
-    SortClause as SortClause,
-)
-from .query import (
-    SortField as SortField,
-)
-from .query import (
-    SortOrder as SortOrder,
-)
 from .store import ProjectionStore as ProjectionStore
