+++ src/logicblocks/event/query/functions.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+from .utilities import Path
+
+
+class Function:
+    pass
+
+
+@dataclass(frozen=True)
+class Similarity(Function):
+    path: Path
+    value: str
