+++ tests/unit/logicblocks/event/query/test_clauses.py
@@ -1,12 +1,4 @@
-from logicblocks.event.projection.store.query import OffsetPagingClause, Path
-
-
-class TestPath:
-    def test_is_not_nested_when_top_level_only(self):
-        assert not Path("a").is_nested()
-
-    def test_is_nested_when_has_sub_levels(self):
-        assert Path("a", "b", "c").is_nested()
+from logicblocks.event.query import OffsetPagingClause
 
 
 class TestOffsetPagingClause:
